fix: indexer

+++ apps/zefyrlab/priv/repo/migrations/20251214195626_create_network_metric_bins.exs
@@ -6,8 +6,8 @@ defmodule Zefyrlab.Repo.Migrations.CreateNetworkMetricBins do
       add :id, :string, null: false
       add :resolution, :string, null: false, primary_key: true
       add :bin, :utc_datetime, null: false, primary_key: true
-      add :volume, :integer, null: false, default: 0
-      add :tvl, :integer, null: false, default: 0
+      add :volume, :bigint, null: false, default: 0
+      add :tvl, :bigint, null: false, default: 0
       add :utilization_ratio, :decimal, null: false, default: 0
 
       timestamps(type: :utc_datetime_usec, null: false)
